Updated resx to X20.0.1, changed Entity.cs member type to correct type

+++ Lib K Relay/Networking/Packets/DataObjects/Entity.cs
@@ -8,12 +8,12 @@ namespace Lib_K_Relay.Networking.Packets.DataObjects
 {
     public class Entity : IDataObject
     {
-        public int ObjectType;
+        public ushort ObjectType;
         public Status Status = new Status();
 
         public IDataObject Read(PacketReader r)
         {
-            ObjectType = (int)r.ReadUInt16();
+            ObjectType = r.ReadUInt16();
             Status.Read(r);
 
             return this;
@@ -21,7 +21,7 @@ namespace Lib_K_Relay.Networking.Packets.DataObjects
 
         public void Write(PacketWriter w)
         {
-            w.Write((ushort)ObjectType);
+            w.Write(ObjectType);
             Status.Write(w);
         }
